requirements.txt
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==2.6.11
pytz==2018.5